+++ src/index.ts
@@ -19,11 +19,13 @@ import {
   InstanceProperties,
   MediaDeviceUpdatePayload,
   QueuedEventListener,
+  QueuedUICallback,
   ReceiveMessage,
   RoomJoinedPayload,
   SendMessage,
   Stored,
   StoredVBState,
+  UICallbackName,
   UserId,
   UserTileType,
   VirtualBackgroundOptions,
@@ -57,6 +59,8 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
   permissionManager = new PermissionManager(this);
 
   queuedEventListeners: QueuedEventListener[] = [];
+
+  queuedUICallbacks: QueuedUICallback[] = [];
 
   constructor(
     options: Partial<InitOptions> = {},
@@ -244,6 +248,39 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
     }
   };
 
+  addUICallback = (name: UICallbackName, listener: (...args: any[]) => void) => {
+    const customEventName = `UICallback_${name}`;
+
+    if (this.connected) {
+      if (!this.listenerCount(customEventName)) {
+        this.sendMessage({ type: 'connectUICallback', data: { name } });
+      }
+    } else {
+      this.queuedUICallbacks.push({
+        operation: 'connectUICallback',
+        name,
+      });
+    }
+
+    this.on(customEventName, listener);
+  };
+
+  removeUICallback = (name: UICallbackName, listener: (...args: any[]) => void) => {
+    const customEventName = `UICallback_${name}`;
+    this.off(customEventName, listener);
+
+    if (this.connected) {
+      if (!this.listenerCount(customEventName)) {
+        this.sendMessage({ type: 'disconnectUICallback', data: { name } });
+      }
+    } else {
+      this.queuedUICallbacks.push({
+        operation: 'disconnectUICallback',
+        name,
+      });
+    }
+  };
+
   private setupInternalEventListeners = () => {
     this.on('userJoined', (event) => {
       const { user, type } = event.data;
@@ -301,6 +338,14 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
 
         this.stored.users[this.stored.userId].dynamicPermissions =
           modifiedPermissions as PermissionTypes[];
+      }
+    });
+
+    this.on('roleChanged', (event) => {
+      const { userId, to } = event.data;
+
+      if (this.stored.users[userId]) {
+        this.stored.users[userId].role = to;
       }
     });
 
@@ -441,6 +486,17 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
         const customEventName = `frameEvent_${eventName}_${target}`;
 
         this._emit(customEventName, JSON.parse(payload));
+        break;
+      }
+
+      case 'UICallback': {
+        const { name } = message.data as any;
+
+        const customEventName = `UICallback_${name}`;
+
+        this._emit(customEventName, {});
+
+        break;
       }
 
       case 'internalMediaDeviceChanged': {
@@ -538,6 +594,7 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
     const payload: ConnectToFramePayload = {
       ...this.roomSettings,
       eventListeners: this.queuedEventListeners,
+      UICallbacks: this.queuedUICallbacks,
     };
 
     this.sendMessage({ type: 'connect', data: payload });
@@ -549,6 +606,7 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
     this.on('connected', () => {
       this.connected = true;
       this.queuedEventListeners = [];
+      this.queuedUICallbacks = [];
 
       clearTimeout(confirmationTimeout);
     });
@@ -866,6 +924,10 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
 
   minimizeContent = () => {
     this.sendMessage({ type: 'minimizeContent' });
+  };
+
+  changeRole = (userId: UserId, role: string) => {
+    this.sendMessage({ type: 'changeRole', data: { userId, role } });
   };
 }